WIP: Getting sound output, but with issues.
I'm not sure if the playback is just not generating sampels in realtime
from the simualation, or what. I seem to have to generate the samples
all in advance before even starting playback, and I have to make sure
that the process stays alive long enough to generate all the samples and
play them back. For some reason it doesn't want to play back samples on
demand.