WIP: MVP carrier compensation working
This should really be part of the algorithm ROM though.
I can either use a block RAM to store the value computed and stored in a
register (which uses resources to decode write messages), or I can
precompute them and store in a block ROM. The second option is much
better.