WIP: Envelopes and fixed sample output bug
Added support for envelopes, but it still needs to be verified.

Also fixed a bug where a voice may not be output, or would output using
the wrong configuration. If it's not clear where, it was in the
"Subsample combination and output" section since the r_SampleBuffer
reset was delivered a clock too late.

The RTL needs major cleanup. The C++ too, really. But at least it's
working better now :)