TEMPORARY HACK: Limiting output to a single voice
I think that this issue where the output is only at ~71% with all output enabled but is near 100% with only a single voice (multiplied by 32) is because of minor phase differences between each of the voices, since they don't all start simultaneously.

I'll just wait for the NOTE_ON and envelope stuff to come along before worrying about this anymore. I'll just use a single voice for now.